+++ lib/datacontext.dart
@@ -1,6 +1,17 @@
 library data_context;
 
+export 'src/models/data-class.dart';
+export 'src/models/data-provider.dart';
+export 'src/models/load-status.dart';
+export 'src/models/data-operation.dart';
 
-export 'src/models/models.dart';
-export 'src/widgets/widgets.dart';
-export 'src/solutions/solutions.dart' show DataSet, DataContext, DataController;
+
+export 'src/widgets/loader-widget.dart';
+export 'src/widgets/fail-widget.dart';
+export 'src/widgets/isnull-widget.dart';
+
+
+
+export 'src/solutions/data-set.dart';
+export 'src/solutions/data-context.dart' show DataContext;
+export 'src/solutions/data-set.dart';