+++ lib/src/models/data-operation.dart
@@ -1,0 +1,3 @@
+enum DataOperation {
+  GET, ADD, UPDATE, REMOVE
+}