+++ lib/src/models/data-provider.dart
@@ -8,6 +8,17 @@ abstract class DataProvider<T extends DataClass> {
 
   abstract final ValueNotifier<LoadStatus> loadStatus;
 
+   /// This list stores data fetched by a data provider or setted by other source
+  abstract List<T> list;
+
+  /// This object stores typed data fetched by a data provider or setted by other source
+  T? get data;
+
+  /// Sets the model to the model passed as parameter and notifier any listener
+  set data (T? data);
+
+  void create(T data);
+
   Future<void> add(T data);
 
   Future<List<T>> get({Map<String, dynamic> filters});