+++ lib/src/solutions/data-context.dart
@@ -1,10 +1,20 @@
-import 'dart:convert';
+import 'package:datacontext/datacontext.dart';
 import 'package:provider/provider.dart';
 import 'package:flutter/material.dart';
 import 'package:http/http.dart';
 
 abstract class DataContext extends ChangeNotifier {
   abstract String origin;
+
+  abstract void Function(Uri, Map<String, String>, Map<String, dynamic>?, DataOperation) onSending;
+
+  abstract void Function(Response) onReceiving;
+
+  DataContext() {
+    DataContextGlobalResources.onSending = onSending;
+    DataContextGlobalResources.onReceiving = onReceiving;
+    DataContextGlobalResources.dataOrigin = origin;
+  }
 
   void setHeader(String key, String value) => DataContextGlobalResources.headers[key] = value;
 
@@ -18,9 +28,11 @@ abstract class DataContext extends ChangeNotifier {
 }
 
 class DataContextGlobalResources {
-  static void Function(Response) bodyParser = (res) => jsonDecode(res.body);
-
   static String dataOrigin = '';
+
+  static void Function(Uri, Map<String, String>, Map<String, dynamic>?, DataOperation) onSending = (a, b, c, d) {};
+
+  static void Function(Response) onReceiving = (a) {};
 
   static Map<String, String> headers = {};
 }